refactor: drop the workarounds objective-z now fixes
Four of the five shapes this app was bent around are gone, because the
defects behind them are fixed and merged upstream:

  - #287 -- the gesture table is an ivar again, `SEL _gestures[PX_KEY_COUNT]`
    inside the `@implementation` rather than a file-scope array.
  - #288 -- `_indicator` carries its protocol again, `id<PXToggleable>`, and
    the three casts that compensated are gone.
  - #288 -- the singleton pointer sits beside its `@implementation`, like
    the other three do, instead of hoisted above the `OZM` blocks.
  - #289 -- the shell command is one `OZM(SHELL_CMD_REGISTER, ..., ^{...})`
    again rather than a named handler plus a bare registration.

The fifth is **permanent**, and now enforced rather than merely advisable.
`PWMOutput` still has no `spec` property, because `GPIOPin` publishes that
selector for a different return type, and dispatch is keyed on the selector
name. #290 turned that from a miscompilation into a located error, so the
shape is diagnosed now -- but it is still not allowed, and cannot become
allowed. Verified by adding the property back and reading the error rather
than assuming.

Also required, and not a workaround: `-cDescription:maxLength:` takes a
`size_t` capacity as of #294, `snprintf`'s own shape. All seven overrides
here were updated. The build tolerated the old `int` through an implicit
conversion, so this is correctness rather than a fix.

`WORKAROUNDS.md` is rewritten around the one entry that remains, with the
removed four recorded against the issue that removed each.

Builds clean for nrf52833dk/nrf52833: FLASH 51.43%, RAM 30.74%. Still not
verified on hardware.

+++ src/GPIOOutput.m
@@ -40,7 +40,7 @@
 	gpio_pin_toggle_dt(super.spec);
 }
 
-- (int)cDescription:(char *)buf maxLength:(int)maxLen
+- (int)cDescription:(char *)buf maxLength:(size_t)maxLen
 {
 	return snprintk(buf, maxLen, "<%s: %s pin %u, active=%d>",
 			"GPIOOutput", super.spec->port->name, super.spec->pin, _active);
